MAINT: Fix to shift main containt according to drawer width"
