+++ frontend/src/Header.js
@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types'
 import React from 'react'
 import { makeStyles } from '@material-ui/core/styles'
 import AppBar from '@material-ui/core/AppBar'
@@ -6,35 +7,14 @@ import Toolbar from '@material-ui/core/Toolbar'
 import Typography from '@material-ui/core/Typography'
 import MenuIcon from '@material-ui/icons/Menu'
 
-import RightDrawer from './RightDrawer'
-
 const useStyles = makeStyles((theme) => ({
-  root: {
-    flexGrow: 1,
-  },
-  appBar: {
-    color: 'white',
-    zIndex: theme.zIndex.drawer + 1,
-    transition: theme.transitions.create(['width', 'margin'], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.leavingScreen,
-    }),
-  },
-  toolbar: {
-    paddingRight: 24, // keep right padding when drawer closed
-  },
   title: {
     flexGrow: 1,
   },
 }))
 
-function Header() {
+function Header({ menuIconClicked }) {
   const classes = useStyles()
-  const [open, setOpen] = React.useState(false)
-
-  const handleDrawerOpen = () => {
-    setOpen(true)
-  }
 
   return (
     <>
@@ -44,13 +24,15 @@ function Header() {
             Random Dish
           </Typography>
           <IconButton color="inherit" aria-label="menu">
-            <MenuIcon onClick={handleDrawerOpen} />
+            <MenuIcon onClick={menuIconClicked} />
           </IconButton>
         </Toolbar>
       </AppBar>
-      <RightDrawer setOpen={setOpen} open={open} />
     </>
   )
+}
+Header.propTypes = {
+  menuIconClicked: PropTypes.func.isRequired,
 }
 
 export default Header
